+++ internal/checks/base_test.go
@@ -596,7 +596,3 @@ func checkErrorUnableToRun(c, name, uri, err string) string {
 func checkErrorTooExpensiveToRun(c, name, uri, err string) string {
 	return fmt.Sprintf("Couldn't run `%s` checks on `%s` Prometheus server at %s because some queries are too expensive: `%s`.", c, name, uri, err)
 }
-
-func checkUnsupported(c, name, uri, path string) string {
-	return fmt.Sprintf("Couldn't run `%s` checks on `%s` Prometheus server at %s because it's unsupported: this server doesn't seem to support `%s` API endpoint.", c, name, uri, path)
-}